Refactor and relocate env commands to info command group
Move python, java, shell, and disk commands from env to info group for better organization. Adjusted imports and logging messages to align with the new structure, ensuring seamless execution of environment-related commands.
