+++ src/hckr/cli/env.py
@@ -119,80 +119,3 @@ def env_list(pattern, ignore_case):
     list_env(pattern, ignore_case)
 
 
-@env.command()
-def python():
-    """
-    This command show all information about your current python environment
-
-    **Example Usage**:
-
-    .. code-block:: shell
-
-        $ hckr env python
-
-
-    **Command Reference**:
-    """
-    # PSuccess(f"Python version: {sys.version}\nPython API version: {sys.api_version}",
-    #          title=f"[yellow]Python {sys.version_info.major}.{sys.version_info.minor}.{sys.version_info.micro}")
-    lib_paths = ''.join(f'  - {path}\n' for path in sys.path)
-    site_package_dirs = ''.join(f'  - {path}\n' for path in site.getsitepackages())
-    info = f"""Python Version: {sys.version}
-Python Executable: {sys.executable}
-Platform Information:
-------------------
-Platform: {sys.platform}
-Python Implementation: {platform.python_implementation()}
-
-Python Library Paths:
----------------------
-{lib_paths}
-Site-Packages Directories:
---------------------------
-{site_package_dirs}
-Python Environment Variables:
------------------------------
-PYTHONPATH: {os.environ.get('PYTHONPATH', 'Not set')}
-PYTHONHOME: {os.environ.get('PYTHONHOME', 'Not set')}
-VIRTUAL_ENV: {os.environ.get('VIRTUAL_ENV', 'Not set')}"""
-    PSuccess(info, title=f"Python {sys.version_info.major}.{sys.version_info.minor}.{sys.version_info.micro}")
-
-
-@env.command()
-def java():
-    try:
-        result = subprocess.run(["java", "-version"], capture_output=True, text=True, check=True)
-        PSuccess(result.stderr, title="[yellow]Java")  # Java version information is typically in stderr
-    except subprocess.CalledProcessError as e:
-        PError(f"Java is not installed or not found in PATH.\n{e}")
-
-
-@env.command()
-def shell():
-    shell_path = os.environ.get('SHELL')
-    shell_name = os.path.basename(shell_path) if shell_path else 'Unknown'
-    PSuccess(f"Current shell: [magenta]{shell_name}[/magenta]\n [yellow]{get_shell_profile()}",
-             title=f"Shell [green]\[{shell_name}]")
-
-
-@env.command()
-def node():
-    try:
-        result = subprocess.run(["node", "--version"], capture_output=True, text=True, check=True)
-        PSuccess(f"Node.js version: {result.stdout.strip()}")
-    except subprocess.CalledProcessError as e:
-        PError(f"Node.js is not installed or not found in PATH.\n{e}")
-
-
-@env.command()
-def disk():
-    total, used, free = shutil.disk_usage("/")
-    info = f"""
-Total: {total // (2 ** 30)}
-GB
-Used: {used // (2 ** 30)}
-GB
-Free: {free // (2 ** 30)}
-GB
-"""
-    PSuccess(info, title=f"Disk Space [green]\[{total // (2 ** 30)} GB]")
