+++ src/hckr/utils/EnvUtils.py
@@ -1,3 +1,4 @@
+import logging
 import os
 import re
 
@@ -27,9 +28,9 @@ def list_env(pattern, ignore_case):
 
 def get_shell_profile(shell=None):
     if shell is None:
-        MessageUtils.info(f"Shell is not provided trying to infer from current shell")
+        logging.info(f"Shell is not provided trying to infer from current shell")
         shell = os.path.basename(os.environ.get('SHELL'))
-        MessageUtils.info(f"Inferred shell as [yellow]{shell}")
+        MessageUtils.info(f"Inferred from current shell as [yellow]{shell}")
     home = os.path.expanduser("~")
 
     if shell == 'bash':
